refactor : ChatCustomRepository 생성하여 리팩토링
