+++ meerket/meerket-domain/src/main/java/org/j1p5/domain/chat/repository/ChatMessageRepository.java
@@ -2,8 +2,9 @@ package org.j1p5.domain.chat.repository;
 
 import org.bson.types.ObjectId;
 import org.j1p5.domain.chat.entity.ChatMessageEntity;
+import org.j1p5.domain.chat.repository.custom.ChatMessageCustomRepository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 
-public interface ChatMessageRepository extends MongoRepository<ChatMessageEntity, ObjectId> {
+public interface ChatMessageRepository extends MongoRepository<ChatMessageEntity, ObjectId>, ChatMessageCustomRepository {
 }
